requirements.txt
matplotlib==3.7.1
numpy==1.23.1
pandas==1.5.3
python-dotenv==1.0.0
seaborn==0.12.2
yfinance==0.2.12
